Install private assets only if requested
Add variables that  governs whether the private dpla_frontend_assets gem
is included for branding-related assets.  This defuaults to false.

Addresses one of the problems raised in
https://github.com/dpla/automation/issues/117

When the gem is included, we write it into the Gemfile, which has been
modified elsewhere (in the app repositories) to not include the gem.

Also:
* Fail build scripts if bundle install fails
* Sneak in some documentation improvements about deployment ssh keys.

+++ ansible/roles/pss/files/build_pss.sh
@@ -1,16 +1,20 @@
 #!/bin/bash
 
 USE_VERSION=$1
+BRANDING=${2:-""}
 export PATH=$HOME/.rbenv/bin:$PATH
 LOGFILE=/tmp/build_pss.log
 
 echo "starting" > $LOGFILE
 
 eval "`rbenv init -`"
-# Start ssh-agent and set environment variables.
-# Work-around for private GitHub repository in Gemfile.
-eval `ssh-agent`
-ssh-add $HOME/git_private_key
+
+if [ "$BRANDING" == "branding" ]; then
+    # Start ssh-agent and set environment variables.
+    # Work-around for private GitHub repository in Gemfile.
+    eval `ssh-agent`
+    ssh-add $HOME/git_private_key
+fi
 
 cd /home/dpla/pss
 
@@ -21,14 +25,19 @@ echo "installing bundle ..." >> $LOGFILE
 
 rm -f Gemfile.lock
 bundle install >> $LOGFILE 2>&1
+if [ $? -ne 0 ]; then
+    exit 1
+fi
 rbenv rehash >> $LOGFILE 2>&1
 
 echo "precompiling assets ..." >> $LOGFILE
 bundle exec rake assets:precompile >> $LOGFILE 2>&1
 
-echo "killing ssh_agent ..." >> $LOGFILE
-# Variable set above by ssh-agent
-kill $SSH_AGENT_PID
+if [ "$BRANDING" == "branding" ]; then
+    echo "killing ssh_agent ..." >> $LOGFILE
+    # Variable set above by ssh-agent
+    kill $SSH_AGENT_PID
+fi
 
 echo "rsync from home to /srv/www/pss ..." >> $LOGFILE
